Tests added; refactored
