+++ script/tests/test_config.py
@@ -1,15 +1,3 @@
-"""Tests for repo file downloading script."""
-
-import json
-from unittest.mock import MagicMock
-
-import pytest
-from aiohttp import ClientSession, web
-from link_extractor import parse_json_content, parse_repo_dir
-
-host: str = 'https://gitea.radium.group/'
-request_headers: str = 'accept: application/json'
-
 all_toml_route: str = '/api/v1/repos/radium/project-configuration/contents/nitpick/all.toml'
 repo_content_all_toml: str = """
 {
@@ -266,67 +254,3 @@ repo_content_nitpick_dir: str = """
   }
 ]
 """
-
-
-async def repo_all_toml_content(request: web.Request) -> web.Response:
-    return web.Response(body=repo_content_all_toml)
-
-
-async def repo_root_content(request: web.Request) -> web.Response:
-    return web.Response(body=repo_content_root_dir)
-
-
-async def repo_nitpick_content(request: web.Request) -> web.Response:
-    return web.Response(body=repo_content_nitpick_dir)
-
-
-def create_app() -> web.Application:
-    app = web.Application()
-    app.router.add_route('GET', all_toml_route, repo_all_toml_content)
-    app.router.add_route('GET', repo_nitpick_dir_route, repo_nitpick_content)
-    app.router.add_route('GET', repo_root_dir_route, repo_root_content)
-    return app
-
-
-@pytest.mark.asyncio()
-async def test_repo_json(aiohttp_client):
-    client = await aiohttp_client(create_app())
-    resp = await client.get(all_toml_route)
-    text = await resp.text()
-    json_data = json.loads(text)
-
-    async with ClientSession() as session:
-        extracted_links_list = await parse_json_content(json_data, session)
-
-    assert extracted_links_list == valid_extracted_links_list_all_toml
-
-
-@pytest.mark.asyncio()
-async def test_parse_nitpick_dir_json_resp(aiohttp_client):
-    client = await aiohttp_client(create_app())
-    resp = await client.get(repo_nitpick_dir_route)
-    text = await resp.text()
-    json_data = json.loads(text)
-
-    extracted_links_list = []
-    async with ClientSession() as session:
-        for json_obj in json_data:
-            link_list = await parse_json_content(json_obj, session)
-            extracted_links_list.extend(link_list)
-
-    assert len(extracted_links_list) == 8
-
-
-@pytest.mark.asyncio()
-async def test_parse_nitpick_dir():
-    extracted_links_list = []
-
-    mock = ClientSession
-    mock.get = MagicMock()
-    mock.get.return_value.__aenter__.return_value.json.return_value = json.loads(repo_content_nitpick_dir)
-
-    async with ClientSession() as session:
-        link_list = await parse_repo_dir(repo_nitpick_dir_route, session)
-        extracted_links_list.extend(link_list)
-
-    assert len(extracted_links_list) == 8
