+++ setup.cfg
@@ -5,6 +5,7 @@ exclude =
     */migrations/,
     venv/,
     env/
+    script/tests/test_config.py
 per-file-ignores = tests/**.py: S101, D103, ANN201
 max-complexity = 10
 spellcheck-targets = comments
